Fragment made
Figuring out how to get it launching the Comment Intent
