+++ QuiRky/app/src/main/java/com/example/quirky/ViewCommentsActivity.java
@@ -66,7 +66,9 @@ public class ViewCommentsActivity extends AppCompatActivity {
         DM = new DatabaseController(FirebaseFirestore.getInstance(), ct);
 
         // Has to be getting it from Fragment
-        String qrCodeId = intent.getStringExtra(MainActivity.EXTRA_MESSAGE); // Needs to be changed
+        String qrCodeId = intent.getStringExtra(ViewQRFragment.EXTRA_MESSAGE);
+        System.out.println(qrCodeId);
+
         commentDataList = new ArrayList<>();
