+++ QuiRky/app/src/main/java/com/example/quirky/ViewQRActivity.java
@@ -8,87 +8,25 @@ import android.view.View;
 import android.widget.Button;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentActivity;
 
 /**
  * This class is of the Activity for when you're clicking on one of the listed items in "Your QR Codes" or
  * "nearby QR code" in "MAP" Activity, as seen on the Project Part 2 Miro.
  * */
-public class ViewQRActivity extends AppCompatActivity {
+public class ViewQRActivity extends FragmentActivity {
 
-//    public static final String EXTRA_MESSAGE = "com.example.QuiRky.MESSAGE";
-
-    private Button Comment;
-    private Button ElseQRCode;
-    private Button SetPrivate;
-    private Button Delete;
-    private Button Location; // But may not be a button because the location description has to
-    // change depending on where the location is.
+    private ViewQRFragment optionFragment;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_view_qr);
 
-        Comment = findViewById(R.id.button_comments);
-        ElseQRCode = findViewById(R.id.button_who_else_has_this_code);
-        SetPrivate = findViewById(R.id.button_set_private);
-        Delete = findViewById(R.id.button_delete);
-        Location = findViewById(R.id.button_location); // would be better if it was a clickable text
+        optionFragment = new ViewQRFragment();
 
-
-        // initialized intent
-        Comment.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                comment();
-            }
-        });
-
-        ElseQRCode.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                elseQRCode();
-            }
-        });
-
-        SetPrivate.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                setPrivate();
-            }
-        });
-
-        Delete.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                delete();
-            }
-        });
-
-        Location.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                location();
-            }
-        });
+        getSupportFragmentManager().beginTransaction()
+                .replace(R.id.qr_fragment, optionFragment).commit();
     }
-
-    /**
-     * Opens the Comment Activity Intent.
-     */
-    public void comment() {
-        Intent intent = new Intent(this, ViewCommentsActivity.class);
-
-        String message = "Sample QR Code ID"; // This needs to be a specific QR code Id.
-        intent.putExtra("comment", message);
-
-        startActivity(intent);
-    }
-
-    // Planning on expanding
-    public void elseQRCode() {     }
-    public void setPrivate() {     }
-    public void delete() {     }
-    public void location() {     }
-
 }
