+++ QuiRky/app/src/main/java/com/example/quirky/ViewQRFragment.java
@@ -6,65 +6,99 @@
 
 package com.example.quirky;
 
+import android.content.Intent;
 import android.os.Bundle;
 
+import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link ViewQRFragment#newInstance} factory method to
+ * Use the {@link ViewQRFragment newInstance} factory method to
  * create an instance of this fragment.
  */
 public class ViewQRFragment extends Fragment {
+    private Button Comment;
+    private Button ElseQRCode;
+    private Button SetPrivate;
+    private Button Delete;
+    private Button Location; // But may not be a button because the location description has to
+    // change depending on where the location is.
 
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
-
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
-
-    public ViewQRFragment() {
-        // Required empty public constructor
-    }
-
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment ViewQRFragment.
-     */
-    // TODO: Rename and change types and number of parameters
-    public static ViewQRFragment newInstance(String param1, String param2) {
-        ViewQRFragment fragment = new ViewQRFragment();
-        Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
-    }
+    View view;
+    public static final String EXTRA_MESSAGE = "com.example.QuiRky.MESSAGE";
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_view_q_r, container, false);
+
+        view =  inflater.inflate(R.layout.fragment_view_q_r, container, false);
+
+//        Comment = view.findViewById(R.id.commentsButtonView);
+//        ElseQRCode = view.findViewById(R.id.elseButtonView);
+//        SetPrivate = view.findViewById(R.id.privateButtonView);
+//        Delete = view.findViewById(R.id.deleteButtonView);
+//        Location = view.findViewById(R.id.locationButtonView);
+//
+//        // initialized intent
+//        Comment.setOnClickListener(new View.OnClickListener() {
+//            @Override
+//            public void onClick(View view) {
+//                comment();
+//            }
+//        });
+//
+//        ElseQRCode.setOnClickListener(new View.OnClickListener() {
+//            @Override
+//            public void onClick(View view) {
+//                elseQRCode();
+//            }
+//        });
+//
+//        SetPrivate.setOnClickListener(new View.OnClickListener() {
+//            @Override
+//            public void onClick(View view) {
+//                setPrivate();
+//            }
+//        });
+//
+//        Delete.setOnClickListener(new View.OnClickListener() {
+//            @Override
+//            public void onClick(View view) {
+//                delete();
+//            }
+//        });
+//
+//        Location.setOnClickListener(new View.OnClickListener() {
+//            @Override
+//            public void onClick(View view) {
+//                location();
+//            }
+//        });
+
+        return view;
     }
+
+
+    public void comment() {
+        Intent intent = new Intent(getActivity(), ViewCommentsActivity.class);
+
+        String message = "Sample QR Code ID"; // This needs to be a specific QR code Id.
+        intent.putExtra("comment", message);
+
+        startActivity(intent);
+
+        System.out.println("Launch Comment");
+    }
+
+    // Planning on expanding
+    public void elseQRCode() {     }
+    public void setPrivate() {     }
+    public void delete() {     }
+    public void location() {     }
 }
